Static IP feedback rework
